Add note
This commit adds a note concerning my inability to bind the <Tab> key in
addition to the <space> key to cycle the connectors. I was able to bind
both Enter keys to the selection logic however and I am not sure why I
cannot get it to work here too.